+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.3.0] - 2025-??-??
+## [0.3.1] - 2025-??-??
+### Added
+- examples/serial.rs
+
+### Changed
+- Fixed bug so that a device when a non-zero serial is used when None is specified for the serial number
+
+## [0.3.0] - 2025-01-31
 ### Added
 - Added [`Device`] as a trait, to allow other device types and easier mocking.  As part of this changed the existing set of functions
 - Similarly made [`BusBuilder`] a trait, to allow creation of [`Bus`] using other device types.